fixed warning

+++ src/media_tools/media_import.c
@@ -534,7 +534,7 @@ GF_Err gf_import_isomedia(GF_MediaImporter *import)
 			samp->DTS -= dts_offset;
 
 			if (samp->nb_pack && duration && (samp->DTS + samp->nb_pack*cdur > duration) ) {
-				u32 nb_samp = (duration - samp->DTS)/cdur;
+				u32 nb_samp = (u32) ( (duration - samp->DTS)/cdur );
 				u32 csize = samp->dataLength / samp->nb_pack;
 				samp->dataLength = csize*nb_samp;
 				samp->nb_pack = nb_samp;
